requirements.txt
google-adk>=1.2.0
google-cloud-aiplatform>=1.80.0
google-cloud-firestore>=2.19.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
mcp>=1.27.0,<2.0.0
pydantic>=2.10.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
